Remove local Gemfile
This was a way for an individual developer to keep
a local Gemfile with optional gems they might like
to use in addition to the ones that are needed for
the project.

I've moved all the useful gems into the canonical
Gemfile and the individual developer can maintain
their own local gems with `gem install` rather use
bundler.